refactor(checkbox): ♻️  move checkbox parts into its own file
