+++ src/checkbox/index.ts
@@ -1,1 +1,5 @@
 export * from "./Checkbox";
+export * from "./CheckboxIcon";
+export * from "./CheckboxInput";
+export * from "./CheckboxLabel";
+export * from "./CheckboxText";
